add sashimi decoder
