+++ configs/bci/baseline_gru.yaml
@@ -1,6 +1,8 @@
 Comment: >
-  # This trains the baseline model
+  # This trains the baseline model (fast_dev_run is a debugging flag)
   python run.py --config-name="baseline_gru" trainer_cfg.fast_dev_run=1
+  # Overfit in 1 batch
+  python run.py --config-name="baseline_gru" trainer_cfg.limit_train_batches=1 trainer_cfg.limit_val_batches=1 trainer_cfg.limit_test_batches=1 trainer_cfg.log_every_n_steps=1
 
 project_name: "speech_bci"
 experiment_name: "baseline_gru"
@@ -15,7 +17,7 @@ common:
 
 data_cfg:
   batchSize: 64
-  num_workers: 16
+  num_workers: 0
   datasetPath: *datasetPath
   # transformations:
   whiteNoiseSD: 0.8  # add white noise
@@ -80,6 +82,10 @@ trainer_cfg:
   accelerator: "auto"
   max_steps: 10000
   log_every_n_steps: 100
+  limit_test_batches:
+  limit_train_batches:
+  limit_val_batches:
+  max_epochs:
 
 
 eval_cfg:
