+++ src/deep_ssm/models/__init__.py
@@ -1,6 +1,7 @@
-from .bci_models import MambaDecoder, GRUDecoder
+from .bci_models import MambaDecoder, GRUDecoder, SashimiDecoder
 
 all_models = {
   "bci_gru" : GRUDecoder,
   "bci_mamba": MambaDecoder,
+  "bci_sashimi": SashimiDecoder
 }
